requirements.txt
google-auth==2.34.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
gspread==6.1.2
icecream==2.1.3